refactor: split framework codegen module
